+++ src/frontend/web_application/server/auth/lib/seal.js
@@ -1,9 +1,0 @@
-import Iron from 'iron';
-
-export function encode(obj, secret, callback) {
-  Iron.seal(obj, secret, Iron.defaults, callback);
-}
-
-export function decode(sealed, secret, callback) {
-  Iron.unseal(sealed, secret, Iron.defaults, callback);
-}